examples/e-commerce-euroelectronics-shopify/requirements.txt
beautifulsoup4==4.14.3
http_client_core==1.5.0
pandas==2.0.3
